do not create dir if it iexist

+++ bootstrap.sh
@@ -4,8 +4,6 @@
 
 # Note: /bin/bash is required for ~/.* expansion in loop below
 
-# Can't use something like 'readlink -e $0' because that doesn't work everywhere
-# And HP doesn't define $PWD in a sudo environment, so we define our own
 case $0 in
     /*|~*)
         SCRIPT_INDIRECT="`dirname $0`"
@@ -19,27 +17,53 @@ esac
 BASEDIR="`(cd \"$SCRIPT_INDIRECT\"; pwd -P)`"
 BASEDIR="${BASEDIR}/dots"
 
-for i in ${BASEDIR}/*; do
-    FILEDIR=`dirname $i`
-    FILE=`basename $i`
-    BASEFILE=$HOME/.$FILE
-    if [ -f $BASEFILE -o -h $BASEFILE ]; then
-        echo "Replacing file: $BASEFILE"
-        rm $BASEFILE
-    else
-        echo "Creating link: $BASEFILE"
-    fi
+function linkRootDots {
+    for i in $1/*; do
+        echo $1
+        if [ -d $i ]; then 
+            return
+        fi
+        FILEDIR=`dirname $i`
+        FILE=`basename $i`
+        BASEFILE=$HOME/.$FILE
+        if [ -f $BASEFILE -o -h $BASEFILE ]; then
+            echo "Replacing file: $BASEFILE"
+            rm $BASEFILE
+        else
+            echo "Creating link: $BASEFILE"
+        fi
 
-    ln -s $i $BASEFILE
-done
+        ln -s $i $BASEFILE
+    done
+    echo "test"
+}
 
-# Make a pass deleting stale links, if any
-for i in ~/.*; do
-    [ ! -h $i ] && continue
+function removeStaleLinks {
+    # Make a pass deleting stale links, if any
+    for i in ~/.*; do
+        [ ! -h $i ] && continue
 
-    # We have a link: Is it stale? If so, delete it ...
-    if [ ! -f $i ]; then
-        echo "Deleting stale link: $i"
-        rm $i
-    fi
-done
+        # We have a link: Is it stale? If so, delete it ...
+        if [ ! -f $i ]; then
+            echo "Deleting stale link: $i"
+            rm $i
+        fi
+    done
+}
+
+function callBootstraps {
+    for i in $1/*; do
+        if [ -d $i ]; then 
+            return
+        fi
+        bootstrap=$i/bootstrap.sh
+        
+        if [[ -f $bootstrap ]]; then
+            bash bootstrap
+        fi        
+    done
+}
+
+linkRootDots $BASEDIR
+callBootstraps $BASEDIR
+removeStaleLinks $1
